requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.0
